+++ src/api/request.js
@@ -1,5 +1,5 @@
 // constants
-import { BASE_URL } from "./constants";
+import { BASE_URL } from "../utils/constants";
