+++ src/services/burger-constructor-slice.js
@@ -3,7 +3,7 @@ import { v4 as uuidv4 } from "uuid";
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 
 // utils
-import request from "../utils/request";
+import request from "../api/request";
